requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
alembic
python-jose
cloudinary
passlib
bcrypt
pydantic-settings
